+++ Models/AdventureWorks/Shift.cs
@@ -12,7 +12,7 @@ namespace resume.Models.AdventureWorks
             EmployeeDepartmentHistories = new HashSet<EmployeeDepartmentHistory>();
         }
 
-        public byte ShiftId { get; set; }
+        public int ShiftId { get; set; }
         public string Name { get; set; }
         public TimeSpan StartTime { get; set; }
         public TimeSpan EndTime { get; set; }
